provide Unpwn.pwned?
This will check the local bloom filter of the top million most common
passwords, and if the offered password passes that check, it will then
run it through the `pwned` gem's check against the haveibeenpwned.com
API to make sure the password hasn't been used anywhere. If both of
those checks pass, or haveibeenpwned is down, it will return `false`.